requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
twilio>=8.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
django-cors-headers>=4.0.0
stripe>=7.0.0
requests>=2.31.0
